requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
asgiref==3.8.1
    # via django
build==1.2.1
    # via pip-tools
bump2version==1.0.1
    # via -r requirements.in
cfgv==3.4.0
    # via pre-commit
click==8.1.7
    # via pip-tools
distlib==0.3.8
    # via virtualenv
django==4.2.13
    # via
    #   django-classy-tags
    #   django-cms
    #   django-formtools
    #   django-sekizai
    #   django-treebeard
    #   djangocms-admin-style
django-classy-tags==4.1.0
    # via
    #   django-cms
    #   django-sekizai
django-cms==4.1.1
    # via -r requirements.in
django-formtools==2.5.1
    # via django-cms
django-sekizai==4.1.0
    # via django-cms
django-treebeard==4.7.1
    # via
    #   -r requirements.in
    #   django-cms
djangocms-admin-style==3.3.1
    # via django-cms
filelock==3.14.0
    # via virtualenv
identify==2.5.36
    # via pre-commit
importlib-metadata==7.1.0
    # via build
nodeenv==1.8.0
    # via pre-commit
packaging==24.0
    # via
    #   build
    #   django-cms
pip-tools==7.4.1
    # via -r requirements.in
platformdirs==4.2.2
    # via virtualenv
pre-commit==3.7.1
    # via -r requirements.in
pyproject-hooks==1.1.0
    # via
    #   build
    #   pip-tools
pyyaml==6.0.1
    # via pre-commit
sqlparse==0.5.0
    # via django
tomli==2.0.1
    # via
    #   build
    #   pip-tools
typing-extensions==4.11.0
    # via asgiref
virtualenv==20.26.2
    # via pre-commit
wheel==0.43.0
    # via
    #   -r requirements.in
    #   pip-tools
zipp==3.18.2
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
